Fix register assignments for hypercall entry.
The existing asm snippets are missing earlyclobber on the output register, so
the compiler may generate incorrect code. I don't think this would happen for
KAFL_HYPERCALL_PT (the natural register allocation would be correct, so we'd
at least need to get inlined somewhere with some particular register allocation
pattern to have issues) but it can happen for KAFL_HYPERCALL_NO_PT even in toy
examples:

original:
 push   %rbx
 mov    $0x8080801f,%r8d
 mov    $0x5658,%eax      <-- compiler thinks it can safely use rax as input
 mov    %rsi,%rcx
 mov    %rdi,%rbx
 mov    %r8,%rax          <-- rax is clobbered before last input is used
 mov    %rax,%rdx         <-- rdx now has wrong value :(
 out    %eax,(%dx)
 mov    %rax,%rax
 pop    %rbx
 ret

A minimal fix would change the outputs to "=&a(_rax)".

fixed:
 push   %rbx
 mov    $0x8080801f,%r8d
 mov    $0x5658,%r9d      <-- compiler now uses a second scratch register :)
 mov    %rsi,%rcx
 mov    %rdi,%rbx
 mov    %r8,%rax
 mov    %r9,%rdx
 out    %eax,(%dx)
 mov    %rax,%rax
 pop    %rbx
 ret

If we instead, use the machine specific constraints to directly assign inputs
to the correct registers, we get better code generation, especially when
inlined.

new:
 push   %rbx
 mov    %rsi,%rcx
 mov    %rdi,%rbx
 mov    $0x8080801f,%eax
 mov    $0x5658,%edx
 out    %eax,(%dx)
 pop    %rbx
 ret

I've only tested this fix for KAFL_HYPERCALL_NO_PT on x86_64, but I've checked
the code generation for the other platforms and it looks correct.