real_network: add packer option to specify client IP:PORT
This commit introduces a new options to allow users to specify a
custom IP address for the client in real network scenarios.
When fuzzing server in a real network scenario, data is sended
via some client in a different network namespace. Introduced options
allow user to specify ip:port of such client.