real_network: transfer data with real network instead of hooks
Data transfer workflow

The Nyx-Net agent replaces the listen function with the underlying listen and then creates a new thread.
The thread is moved to a separate network namespace and acts as a client. Its job is to receive data from the fuzzer and transfer it.
The thread connects to the server using the real connect function and then idles, waiting for the fuzzer's notification to transfer data.
Server calls hooked accept, TCP handshake is established.
Server calls hooked recv function that does the following:
    Invokes memory snapshot, capturing the server and client after the TCP handshake.
    Gets data from the fuzzer as usual, using the handle_next_packet() function.
    Stores received data in a shared variable and notify the thread.
    Calls real recv.
When the thread receives a notification it sends the data stored in a shared variable using the real write function and then terminates.
Client will eventually get the data from its hooked recv function invocation and starts processing it.
After the data has been processed, the memory snapshot is restored and the process starts again.