enable userspace port I/O for hypercalls in guest
Use ioperm() to explicitly allow port I/O to the port used for Nyx
hypercalls. With this, the host kernel no longer needs to enable
enable_vmware_backdoor=y.

Benefits:

 - It avoids needing to change system-wide host configuration.
 - It lets KVM avoid intercepting guest #GP and decoding every instruction
   that hits #GP.

Downsides:

 - Some stuff in the guest kernel will hit slowpaths; in particular,
   the syscall return path will probably run a few instructions more
   (tss_update_io_bitmap()).

Switching between multiple tasks with separately-created ioperm()
bitmasks would probably be slow because we'd hit the memcpy() in
tss_copy_io_bitmap(), but as long as we just set up a single ioperm()
bitmask in init, that shouldn't happen.